pull the cfg logging and dict printing out into print_utils.py
